+++ CHANGELOG.md
@@ -40,6 +40,8 @@ The changes listed in this file are categorised as follows:
 
 ### Changed
 
+* ([#90](https://github.com/iiasa/climate-assessment/pull/90])) Update CO2 sectors for 
+  newer common-definitions
 * ([#55](https://github.com/iiasa/climate-assessment/pull/55])) Switched to using poetry
   for environment management
 * ([#55](https://github.com/iiasa/climate-assessment/pull/55])) Include a `poetry.lock`